requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
Jinja2>=3.1.3
python-dateutil>=2.9.0.post0
# dev
pytest>=8.0.0
tzdata>=2024.1

